REV: mutiple view  mode of process definitions  and refacting for baisc and advanced mode      and some formatter issue fix
