+++ jbpm-console-ng-generic/jbpm-console-ng-generic-client/src/main/resources/org/jbpm/console/ng/gc/client/i18n/Constants.properties
@@ -9,3 +9,4 @@ Month=Month
 Today=Today
 Previous=Previous
 Next=Next
+
