+++ jbpm-console-ng-process-runtime/jbpm-console-ng-process-runtime-client/src/main/java/org/jbpm/console/ng/pr/client/editors/definition/details/advance/AdvancedProcessDefDetailsViewImpl.html
@@ -13,43 +13,36 @@
           <label data-field="processNameText"/>
         </div>
       </div>
-     
-      
       <div class="control-group">
         <strong><label class="control-label" for="deploymentIdText" data-field="deploymentIdLabel"></label></strong>
         <div class="controls readonly">
           <label data-field="deploymentIdText"/>
         </div>
       </div>
-      
       <div class="control-group">
         <strong><label class="control-label" for="humanTasksListBox" data-field="humanTasksListLabel"></label></strong>
         <div class="controls readonly">
           <label data-field="humanTasksListBox" />
         </div>
       </div>
-      
       <div class="control-group">
         <strong><label class="control-label" for="nroOfHumanTasksText" data-field="nroOfHumanTasksLabel"></label></strong>
         <div class="controls readonly">
           <label data-field="nroOfHumanTasksText"/>
         </div>
       </div>
-        
       <div class="control-group">
         <strong><label class="control-label" for="usersGroupsListBox" data-field="usersGroupsListLabel"></label></strong>
         <div class="controls readonly">
           <label data-field="usersGroupsListBox" />
         </div>
       </div>
-      
       <div class="control-group">
         <strong><label class="control-label" for="subprocessListBox" data-field="subprocessListLabel"></label></strong>
         <div class="controls readonly">
           <label data-field="subprocessListBox"/>
         </div>
       </div>
-      
       <div class="control-group">
         <strong><label class="control-label" for="processDataListBox" data-field="processDataListLabel"></label></strong>
         <div class="controls readonly">
@@ -62,7 +55,6 @@
           <label data-field="processServicesListBox"/>
         </div>
       </div>
-
     </div>
   </div>
 </div>
