+++ jbpm-console-ng-process-runtime/jbpm-console-ng-process-runtime-client/src/main/java/org/jbpm/console/ng/pr/client/editors/definition/details/basic/BasicProcessDefDetailsViewImpl.html
@@ -1,0 +1,24 @@
+<div style="margin-top:20px" class="container-fluid">
+  <div class="row-fluid">
+    <div class="form-horizontal">
+      <div class="control-group">
+        <strong><label class="control-label" for="processIdText" data-field="processIdLabel"></label></strong>
+        <div class="controls readonly">
+          <label data-field="processIdText" />
+        </div>
+      </div>
+      <div class="control-group">
+        <strong><label class="control-label" for="processNameText" data-field="processNameLabel"></label></strong>
+        <div class="controls readonly">
+          <label data-field="processNameText"/>
+        </div>
+      </div>
+      <div class="control-group">
+        <strong><label class="control-label" for="deploymentIdText" data-field="deploymentIdLabel"></label></strong>
+        <div class="controls readonly">
+          <label data-field="deploymentIdText"/>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
