+++ jbpm-console-ng-process-runtime/jbpm-console-ng-process-runtime-client/src/main/java/org/jbpm/console/ng/pr/client/editors/definition/details/multi/advance/AdvancedProcessDefDetailsMultiViewImpl.ui.xml
@@ -27,9 +27,7 @@
 
   <g:FlowPanel addStyleNames="{style.dataGridContainer}">
     <g:TabLayoutPanel ui:field="tabsContainer" barUnit="EM" barHeight="3">
-
     </g:TabLayoutPanel>
-
   </g:FlowPanel>
 
 </ui:UiBinder>
